refactor(profile/reports/components/listening-days-chart): move formatter to `/helpers`
