+++ app/profile/[id]/reports/components/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from './chart-tooltip-content-formatter'
