+++ app/profile/[id]/reports/components/listening-days-chart.tsx
@@ -1,8 +1,9 @@
 'use client'
 
 import prettyMilliseconds from 'pretty-ms'
-import type { ComponentProps, CSSProperties } from 'react'
 import { Bar, BarChart, CartesianGrid, XAxis, YAxis } from 'recharts'
+
+import { chartTooltipContentFormatter } from './helpers'
 
 import { StatsMeasurement } from '@app/api/enums'
 import type { ChartConfig } from '@app/components/ui/chart'
@@ -76,42 +77,6 @@ function ListeningDaysChart({
     },
   } satisfies ChartConfig
 
-  type ChartTooltipContentFormatter = ComponentProps<
-    typeof ChartTooltipContent
-  >['formatter']
-
-  const formatter: ChartTooltipContentFormatter = (value, name, item) => {
-    const formattedValue =
-      measurement === StatsMeasurement.PLAYS
-        ? `${value} plays`
-        : prettyMilliseconds(+value, { unitCount: 2 })
-
-    return (
-      <>
-        <div
-          className="shrink-0 rounded-[2px] border-[--color-border] bg-[--color-bg] h-2.5 w-2.5"
-          style={
-            {
-              '--color-bg': item.color,
-              '--color-border': item.color,
-            } as CSSProperties
-          }
-        />
-
-        <div className="flex flex-1 gap-1 justify-between leading-none items-end">
-          <div className="grid gap-1.5">
-            <span className="text-muted-foreground">{item.name}</span>
-          </div>
-          {item.value && (
-            <span className="font-mono font-medium tabular-nums text-foreground">
-              {formattedValue}
-            </span>
-          )}
-        </div>
-      </>
-    )
-  }
-
   const tickFormatter = (value: number) =>
     measurement === StatsMeasurement.PLAYS
       ? value + ''
@@ -135,7 +100,10 @@ function ListeningDaysChart({
           tickFormatter={tickFormatter}
         />
         <CartesianGrid vertical={false} />
-        <ChartTooltip content={<ChartTooltipContent />} formatter={formatter} />
+        <ChartTooltip
+          content={<ChartTooltipContent />}
+          formatter={chartTooltipContentFormatter(measurement)}
+        />
         <ChartLegend content={<ChartLegendContent />} />
         <Bar dataKey="thisWeek" fill="var(--color-thisWeek)" radius={4} />
         <Bar dataKey="lastWeek" fill="var(--color-lastWeek)" radius={4} />
